// src/app/core/models/questionnaire-naep.model.ts
export class QuestionnaireNaep {
    id: number;
    customerName: string;
    dialCode: string;
    phoneNumer: string;
    phoneNumerCustomer: string;
    recruiterName: string;
    recruiterId: string;
    teamManagerName: string;
    teamManagerId: string;
    startTime: string;
    endTime: string;
    questionnaireNaepApprove;
    uuid : string;
    updateTime: string;
    is_answer: boolean;
    
}